requirements.txt
slack_sdk==3.27.1
requests>=2.31.0
python-dotenv>=1.0.1
typer>=0.12.0
openai>=1.40.0
pydantic>=2.9.0
tenacity>=8.4.1
PyGithub>=2.1.1
# SaaS dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
